Stabilize Security acceptance harness

+++ src/Apps/SecuritySQLite.AcceptanceTests/SecurityWebApplicationFactoryExtensions.cs
@@ -3,27 +3,47 @@ using cCoder.Security.Data.Models;
 using cCoder.Security.Exposures;
 using cCoder.Security.Objects.Entities;
 using Microsoft.AspNetCore.Mvc.Testing;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Security.AcceptanceTests;
 
 public static class SecurityWebApplicationFactoryExtensions
 {
+    private static readonly object SetupLock = new();
+    private static bool acceptanceEnvironmentConfigured = false;
     private static bool setupComplete = false;
 
     public static void EnsureDatabasesAreSetupForTesting(this WebApplicationFactory<AcceptanceHost> appFactory)
     {
-        lock (appFactory)
+        lock (SetupLock)
         {
             if (setupComplete)
                 return;
 
-            setupComplete = true;
+            ConfigureAcceptanceEnvironment();
 
             using IServiceScope scope = appFactory.Services.CreateScope();
             IServiceProvider scopedServices = scope.ServiceProvider;
             EnsureSSOSetupForTesting(scopedServices).AsTask().Wait();
+            setupComplete = true;
         }
+    }
+
+    private static void ConfigureAcceptanceEnvironment()
+    {
+        if (acceptanceEnvironmentConfigured)
+            return;
+
+        if (typeof(AcceptanceHost).Namespace == "SecurityMSSQL")
+        {
+            string databaseName = $"SSOAcceptanceTests_{Environment.ProcessId}";
+            Environment.SetEnvironmentVariable(
+                "ENV_ConnectionStrings__SSO",
+                $"Data Source=(localdb)\\MSSQLLocalDB;Initial Catalog={databaseName};MultipleActiveResultSets=True;Trusted_Connection=True;Trust Server Certificate=true");
+        }
+
+        acceptanceEnvironmentConfigured = true;
     }
 
     private static async ValueTask EnsureSSOSetupForTesting(IServiceProvider scopedServices)
@@ -34,7 +54,9 @@ public static class SecurityWebApplicationFactoryExtensions
         using cCoder.Security.Data.EF.SecurityDbContext db =
             scopedServices.GetRequiredService<ISecurityDbContextFactory>().CreateDbContext();
 
-        db.Database.EnsureDeleted();
+        if (db.Database.ProviderName?.Contains("SqlServer", StringComparison.OrdinalIgnoreCase) != true)
+            db.Database.EnsureDeleted();
+
         db.Migrate();
 
         await SetupTestUser(tenantManager);
